requirements.txt
allure-pytest==2.13.0
allure-python-commons==2.13.0
astroid==2.15.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
distlib==0.3.6
exceptiongroup==1.1.1
execnet==1.9.0
filelock==3.10.6
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.0
platformdirs==3.1.1
playwright==1.32.1
pluggy==1.0.0
pycparser==2.21
pyee==9.0.4
pylint==2.17.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.2
pytest-base-url==2.0.0
pytest-env==0.8.1
pytest-playwright==0.3.2
pytest-xdist==3.2.1
python-dotenv==1.0.0
python-slugify==6.1.2
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomlkit==0.11.6
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wrapt==1.15.0
wsproto==1.2.0
